# apps/indigo-api-gateway/application/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API gateway for Indigo Server.",
        "title": "Indigo Server",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/organization/admin_api_keys": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a paginated list of all admin API keys for the authenticated organization.",
                "tags": [
                    "Administration API"
                ],
                "summary": "List Admin API Keys",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "The maximum number of items to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "A cursor for use in pagination. The ID of the last object from the previous page",
                        "name": "after",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the list of admin API keys",
                        "schema": {
                            "$ref": "#/definitions/app_interfaces_http_routes_v1_organization.AdminApiKeyListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing API key",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new admin API key for an organization. Requires a valid admin API key in the Authorization header.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Administration API"
                ],
                "summary": "Create Admin API Key",
                "parameters": [
                    {
                        "description": "API key creation request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app_interfaces_http_routes_v1_organization.CreateOrganizationAdminAPIKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created admin API key",
                        "schema": {
                            "$ref": "#/definitions/app_interfaces_http_routes_v1_organization.OrganizationAdminAPIKeyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid payload",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing API key",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/organization/projects": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a paginated list of all projects for the authenticated organization.",
                "tags": [
                    "Administration API"
                ],
                "summary": "List Projects",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "The maximum number of items to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "A cursor for use in pagination. The ID of the last object from the previous page",
                        "name": "after",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Whether to include archived projects.",
                        "name": "include_archived",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the list of projects",
                        "schema": {
                            "$ref": "#/definitions/app_interfaces_http_routes_v1_organization_projects.ProjectListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing API key",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new project for an organization.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Administration API"
                ],
                "summary": "Create Project",
                "parameters": [
                    {
                        "description": "Project creation request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app_interfaces_http_routes_v1_organization_projects.CreateProjectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created project",
                        "schema": {
                            "$ref": "#/definitions/app_interfaces_http_routes_v1_organization_projects.ProjectResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid payload",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing API key",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app_interfaces_http_routes_v1_organization.AdminApiKeyListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app_interfaces_http_routes_v1_organization.OrganizationAdminAPIKeyResponse"
                    }
                },
                "first_id": {
                    "type": "string"
                },
                "has_more": {
                    "type": "boolean"
                },
                "last_id": {
                    "type": "string"
                },
                "object": {
                    "type": "string",
                    "example": "list"
                }
            }
        },
        "app_interfaces_http_routes_v1_organization.CreateOrganizationAdminAPIKeyRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "My Admin API Key"
                }
            }
        },
        "app_interfaces_http_routes_v1_organization.OrganizationAdminAPIKeyResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer",
                    "example": 1698765432
                },
                "id": {
                    "type": "string",
                    "example": "key_1234567890"
                },
                "last_used_at": {
                    "type": "integer",
                    "example": 1698765432
                },
                "name": {
                    "type": "string",
                    "example": "My Admin API Key"
                },
                "object": {
                    "type": "string",
                    "example": "api_key"
                },
                "owner": {
                    "$ref": "#/definitions/app_interfaces_http_routes_v1_organization.Owner"
                },
                "redacted_value": {
                    "type": "string",
                    "example": "sk-...abcd"
                },
                "value": {
                    "type": "string",
                    "example": "sk-abcdef1234567890"
                }
            }
        },
        "app_interfaces_http_routes_v1_organization.Owner": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer",
                    "example": 1698765432
                },
                "id": {
                    "type": "string",
                    "example": "user_1234567890"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "object": {
                    "type": "string",
                    "example": "user"
                },
                "role": {
                    "type": "string",
                    "example": "admin"
                },
                "type": {
                    "type": "string",
                    "example": "user"
                }
            }
        },
...
        "app_interfaces_http_routes_v1_organization_projects.CreateProjectRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "New AI Project"
                }
            }
        },
        "app_interfaces_http_routes_v1_organization_projects.ProjectListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app_interfaces_http_routes_v1_organization_projects.ProjectResponse"
                    }
                },
                "first_id": {
                    "type": "string"
                },
                "has_more": {
                    "type": "boolean"
                },
                "last_id": {
                    "type": "string"
                },
                "object": {
                    "type": "string",
                    "example": "list"
                }
            }
        },
        "app_interfaces_http_routes_v1_organization_projects.ProjectResponse": {
            "type": "object",
            "properties": {
                "archived_at": {
                    "type": "integer",
                    "example": 1698765432
                },
                "created_at": {
                    "type": "integer",
                    "example": 1698765432
                },
                "id": {
                    "type": "string",
                    "example": "proj_1234567890"
                },
                "name": {
                    "type": "string",
                    "example": "My First Project"
                },
                "object": {
                    "type": "string",
                    "example": "project"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}